Update rb.py
Notes for Ben: It took me a bit, but I got the application to parse our the headers as well as the CSP. Changes the colors to something more readable (Green), Labelled the headers, and basically was prettying it up. 

Would love your thoughts on how we can improve it and also what other apps we should add. I did fix the output to log issue vs. what's displayed on the terminal. It's not pretty, but it works. Basically I f.write and then go back and print(). That's where I left off with it. It's connecting to burp now and I think we're close to just tweaking it for basic webapp reconaissance. 

One of my friends suggested using a wordpress app as well.